remove ipfs network code
