+++ cmd/boost/retrieve_retrieval.go
@@ -20,10 +20,7 @@ import (
 	"github.com/filecoin-project/go-fil-markets/storagemarket"
 	"github.com/filecoin-project/go-state-types/big"
 	"github.com/filecoin-project/lotus/chain/types"
-	"github.com/ipfs/go-blockservice"
 	"github.com/ipfs/go-cid"
-	ipldformat "github.com/ipfs/go-ipld-format"
-	"github.com/ipfs/go-merkledag"
 	"github.com/ipld/go-ipld-prime"
 	"golang.org/x/term"
 	"golang.org/x/xerrors"
@@ -33,104 +30,6 @@ import (
 // over a specific network
 type RetrievalAttempt interface {
 	Retrieve(context.Context, *Node) (RetrievalStats, error)
-}
-
-type IPFSRetrievalAttempt struct {
-	Cid cid.Cid
-}
-
-func (attempt *IPFSRetrievalAttempt) Retrieve(ctx context.Context, node *Node) (RetrievalStats, error) {
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	log.Info("Searching IPFS for CID...")
-
-	providers := node.DHT.FindProvidersAsync(ctx, attempt.Cid, 0)
-
-	// Ready will be true if we connected to at least one provider, false if no
-	// miners successfully connected
-	ready := make(chan bool, 1)
-	go func() {
-		for {
-			select {
-			case provider, ok := <-providers:
-				if !ok {
-					ready <- false
-					return
-				}
-
-				// If no addresses are listed for the provider, we should just
-				// skip it
-				if len(provider.Addrs) == 0 {
-					log.Debugf("Skipping IPFS provider with no addresses %s", provider.ID)
-					continue
-				}
-
-				log.Infof("Connected to IPFS provider %s", provider.ID)
-				ready <- true
-			case <-ctx.Done():
-				return
-			}
-		}
-	}()
-
-	select {
-	// TODO: also add connection timeout
-	case <-ctx.Done():
-		return nil, ctx.Err()
-	case ready := <-ready:
-		if !ready {
-			return nil, fmt.Errorf("couldn't find CID")
-		}
-	}
-
-	// If we were able to connect to at least one of the providers, go ahead
-	// with the retrieval
-
-	var progressLk sync.Mutex
-	var bytesRetrieved uint64 = 0
-	startTime := time.Now()
-
-	log.Info("Starting retrieval")
-
-	bserv := blockservice.New(node.Blockstore, node.Bitswap)
-	dserv := merkledag.NewDAGService(bserv)
-	//dsess := dserv.Session(ctx)
-
-	cset := cid.NewSet()
-	if err := merkledag.Walk(ctx, func(ctx context.Context, c cid.Cid) ([]*ipldformat.Link, error) {
-		node, err := dserv.Get(ctx, c)
-		if err != nil {
-			return nil, err
-		}
-
-		// Only count leaf nodes toward the total size
-		if len(node.Links()) == 0 {
-			progressLk.Lock()
-			nodeSize, err := node.Size()
-			if err != nil {
-				nodeSize = 0
-			}
-			bytesRetrieved += nodeSize
-			printProgress(bytesRetrieved)
-			progressLk.Unlock()
-		}
-
-		if c.Type() == cid.Raw {
-			return nil, nil
-		}
-
-		return node.Links(), nil
-	}, attempt.Cid, cset.Visit, merkledag.Concurrent()); err != nil {
-		return nil, err
-	}
-
-	log.Info("IPFS retrieval succeeded")
-
-	return &IPFSRetrievalStats{
-		ByteSize: bytesRetrieved,
-		Duration: time.Since(startTime),
-	}, nil
 }
 
 type FILRetrievalAttempt struct {
@@ -149,9 +48,6 @@ func (attempt *FILRetrievalAttempt) Retrieve(ctx context.Context, node *Node) (R
 		log.Info("No miners were provided, will not attempt FIL retrieval")
 		return nil, xerrors.Errorf("retrieval failed: no miners were provided")
 	}
-
-	// If IPFS retrieval was unavailable, do a full FIL retrieval. Start with
-	// querying all the candidates for sorting.
 
 	log.Info("Querying FIL retrieval candidates...")
 
@@ -324,25 +220,8 @@ func (stats *FILRetrievalStats) GetAverageBytesPerSecond() uint64 {
 	return stats.AverageSpeed
 }
 
-type IPFSRetrievalStats struct {
-	ByteSize uint64
-	Duration time.Duration
-}
-
-func (stats *IPFSRetrievalStats) GetByteSize() uint64 {
-	return stats.ByteSize
-}
-
-func (stats *IPFSRetrievalStats) GetDuration() time.Duration {
-	return stats.Duration
-}
-
-func (stats *IPFSRetrievalStats) GetAverageBytesPerSecond() uint64 {
-	return uint64(float64(stats.ByteSize) / stats.Duration.Seconds())
-}
-
 // Takes a list of network configs to attempt to retrieve from, in order. Valid
-// structs for the interface: IPFSRetrievalConfig, FILRetrievalConfig
+// structs for the interface: FILRetrievalConfig
 func (node *Node) RetrieveFromBestCandidate(
 	ctx context.Context,
 	attempts []RetrievalAttempt,
@@ -476,21 +355,10 @@ Peer:          %v
 			stats.NumPayments,
 			stats.Peer,
 		)
-	case *IPFSRetrievalStats:
-		fmt.Printf(`RETRIEVAL STATS (IPFS)
------
-Size:          %v (%v)
-Duration:      %v
-Average Speed: %v
-`,
-			stats.ByteSize, humanize.IBytes(stats.ByteSize),
-			stats.Duration,
-			stats.GetAverageBytesPerSecond(),
-		)
 	}
 }
 
-func printQueryResponse(query *retrievalmarket.QueryResponse, availableOnIPFS bool) {
+func printQueryResponse(query *retrievalmarket.QueryResponse) {
 	var status string
 	switch query.Status {
 	case retrievalmarket.QueryResponseAvailable:
@@ -541,9 +409,5 @@ Max Payment Interval Increase: %v (%v)
 
 	if query.Message != "" {
 		fmt.Printf("Message: %v\n", query.Message)
-	}
-
-	if availableOnIPFS {
-		fmt.Printf("-----\nAvaiable on IPFS")
 	}
 }
